rename TaskDefinition to CommandTaskDefinition
